split pteams tests file in requests
